backend/requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
